Test automation expansion
This fully encapsulates the test flow into Python. Use of tio has been removed
and replaced entirely by pyserial, removing the complication of an external
tool.

TODO: The test firmware itself needs a clear delineator when it is complete
that we can look for. As it stands, tests just sort of run and complete with
inconsistent syntax, and eventually we run out of tests and get stuck looking
at timer ISR test leftovers.

As it stands the script will time out after 600 seconds. All test output is
logged to a file within the working directory. The script also serves
as a tool to manipulate the GPIOs controlling boot functionality, and
can be configured to target different GPIOs if needed.